refactor: fix variables
